+++ app/src/main/java/com/moduro/barrier_free_app/presentation/mypage/navigation/MypageNavGraph.kt
@@ -2,6 +2,7 @@ package com.moduro.barrier_free_app.presentation.mypage.navigation
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
+import com.moduro.barrier_free_app.presentation.mypage.screen.FavoritePlaceScreen
 import com.moduro.barrier_free_app.presentation.mypage.screen.MypageRoute
 import com.moduro.barrier_free_app.presentation.mypage.screen.ProfileSettingScreen
 
@@ -17,5 +18,8 @@ fun NavGraphBuilder.mypageNavGraph(
             onBackClick = { navigator.navController.popBackStack() },
             onNavigateToMypage = { navigator.navigateToMypage() }
         )
+    }
+    composable("favorite_place"){
+        FavoritePlaceScreen(onBackClick = {navigator.navController.popBackStack()})
     }
 }
