+++ app/src/main/java/com/moduro/barrier_free_app/presentation/mypage/navigation/MypageNavigator.kt
@@ -11,5 +11,8 @@ class MypageNavigator(
     fun navigateToMypage() {
         navController.navigate("mypage")
     }
+    fun navigateToFavoritePlace() {
+        navController.navigate("favorite_place")
+    }
 
 }
